Solve "Day 5: Binary Boarding"
I felt like doing something different this time.

Part 1: no code
---------------

The first part was solved without any code. Instead, the puzzle input was
investigated manually for the highest seat id, by filtering out seats not relevant.
This is done by looking for upper half turns ("B" and "R") at the start of the
line and then stepping forward, following this thinking:

 1. If the next character is a "B" (or "R"), continue.
 2. Otherwise, go for "F" (or "L").

At characters 8-10, "R" and "L" was used instead of "B" and "F"

For this, the editor of choice search function can be used, preferrably
in regexp mode. This was the progress for the provided puzzle input:

 - Starting point: all 908 boarding passes remain
 - `^B`: 441 boarding passes remain
 - `^BB`: 186 boarding passes remain
 - `^BBB`: 58 boarding passes remain
 - `^BBBB`: 0 remain, instead using `^BBBF`
 - `^BBBFB`: 26 boarding passes remain
 - `^BBBFBB`: 10 boarding passes remain
 - `^BBBFBBB`: 2 boarding passes remain

Remaining boarding passes, e.q. the highest seat ids:

 - BBBFBBBLLR (highest)
 - BBBFBBBLLL (lower, filtred out)

 After this, some manual binary searching was made to find the value.

 Eventually, I also wrote the code for it. Since filtering is done, Part
 1 is a fast program to execute since all seat ids are not calculated,
 only the highest.

 This code is then reused in Part 2.

Part 2: print debugging to the resque
-------------------------------------

 For part 2, a manual investigation of the remaining seats was required to
 separate the seat from the missing seats. print debugging was used here
 for this, and the output looked like this:

 frozenset({0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,
 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34,
 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 615, 954, 955, 956, 957, 958,
 959, 960, 961, 962, 963, 964, 965, 966, 967, 968, 969, 970, 971, 972,
 973, 974, 975, 976, 977, 978, 979, 980, 981, 982, 983, 984, 985, 986,
 987, 988, 989, 990, 991, 992, 993, 994, 995, 996, 997, 998, 999, 1000,
 1001, 1002, 1003, 1004, 1005, 1006, 1007, 1008, 1009, 1010, 1011, 1012,
 1013, 1014, 1015, 1016, 1017, 1018, 1019, 1020, 1021, 1022, 1023})

Note that `615` is standing out, and that the other values are at the
start and the end of the range (0-1024). This is the answer based on th
provided puzzle input.