Solve "Day 6: Custom Customs"
Solution of Part 1 did not require the parsed puzzle input to recognize
the groups, so I was tempted to add `"".join(groups)` to the puzzle input
parser.

But I had a feeling Part 2 would require groups acknowlegements, so therefore,
`"".join(group)` was not added to the Puzzle input parsing.

Turned out I was right, since Part 2 required the group length to be
solved.