Solve "Day 16: Ticket Translation"
Ugh, so much input parsing for this one.

Part 2 was ultra-heavy to solve, and without looking into the subreddit
I would not have known that the columns could match more than 1 range.

It was hard to test this, since the example data did not include any
departure ranges.

The old AOC trap of having an example not showcasing problems that will bust
yer head with the real puzzle data tricked me once again.